+++ test/coverlet.core.coverage.tests/Coverage/CoverageTests.DoesNotReturn.cs
@@ -52,6 +52,7 @@ namespace Coverlet.CoreCoverage.Tests
     public void If_DoesNotReturnAttribute_InstrumentsCorrect()
     {
       Assert.SkipWhen(TestEnvironment.IsVisualStudio, TestEnvironment.VisualStudioSkipMessage);
+      Assert.SkipWhen(TestEnvironment.HasInteractiveStdin, TestEnvironment.InteractiveStdinSkipMessage);
       string path = Path.GetTempFileName();
       try
       {
@@ -84,6 +85,7 @@ namespace Coverlet.CoreCoverage.Tests
     public void Switch_DoesNotReturnAttribute_InstrumentsCorrect()
     {
       Assert.SkipWhen(TestEnvironment.IsVisualStudio, TestEnvironment.VisualStudioSkipMessage);
+      Assert.SkipWhen(TestEnvironment.HasInteractiveStdin, TestEnvironment.InteractiveStdinSkipMessage);
       string path = Path.GetTempFileName();
       try
       {
@@ -116,6 +118,7 @@ namespace Coverlet.CoreCoverage.Tests
     public void Subtle_DoesNotReturnAttribute_InstrumentsCorrect()
     {
       Assert.SkipWhen(TestEnvironment.IsVisualStudio, TestEnvironment.VisualStudioSkipMessage);
+      Assert.SkipWhen(TestEnvironment.HasInteractiveStdin, TestEnvironment.InteractiveStdinSkipMessage);
       string path = Path.GetTempFileName();
       try
       {
@@ -148,6 +151,7 @@ namespace Coverlet.CoreCoverage.Tests
     public void UnreachableBranch_DoesNotReturnAttribute_InstrumentsCorrect()
     {
       Assert.SkipWhen(TestEnvironment.IsVisualStudio, TestEnvironment.VisualStudioSkipMessage);
+      Assert.SkipWhen(TestEnvironment.HasInteractiveStdin, TestEnvironment.InteractiveStdinSkipMessage);
       string path = Path.GetTempFileName();
       try
       {